samba and swat should both depend directly on update-inetd, because update-inetd is the package that provides the update-inetd (some implementations of inet-superserver do *not* depend on update-inetd!).
swat also needs to depend on openbsd-inetd | inet-superserver, because the
superserver is basic to swat's functionality; for samba, it's a Suggests:
only because inetd is not used by this package by default.

Also make the same changes to debian/control.in, so they don't get
overwritten later by a refresh.



git-svn-id: svn://svn.debian.org/svn/pkg-samba/trunk/samba@1412 fc4039ab-9d04-0410-8cac-899223bdd6b0